+++ model_fetch_firestore/lib/model_fetch_firestore.dart
@@ -1,10 +1,8 @@
 export 'src/collection/firestore_frozen_lazy_load_bloc.dart';
 export 'src/collection/firestore_lazy_load_bloc.dart';
 export 'src/collection/firestore_live_single_document_collection_bloc.dart';
-
 export 'src/loader_factories/abstract.dart';
-
+export 'src/macros/firestore_model.dart';
 export 'src/model_by_filter/firestore_live_by_filter_bloc.dart';
 export 'src/model_by_id/firestore_live_by_id_bloc.dart';
-
 export 'src/query_builder.dart';